+++ clutter/clutter/clutter-box-layout.c
@@ -519,12 +519,12 @@ compare_gap (gconstpointer p1,
   const guint *c1 = p1;
   const guint *c2 = p2;
 
-  const gint d1 = MAX (sizes[*c1].natural_size -
-                       sizes[*c1].minimum_size,
-                       0);
-  const gint d2 = MAX (sizes[*c2].natural_size -
-                       sizes[*c2].minimum_size,
-                       0);
+  const int d1 = (int) MAX (sizes[*c1].natural_size -
+                            sizes[*c1].minimum_size,
+                            0);
+  const int d2 = (int) MAX (sizes[*c2].natural_size -
+                            sizes[*c2].minimum_size,
+                            0);
 
   gint delta = (d2 - d1);
 
@@ -601,7 +601,7 @@ distribute_natural_allocation (float          extra_space,
        * Sort order and reducing remaining space by assigned space
        * ensures that space is distributed equally.
        */
-      float glue = (extra_space + i) / (i + 1.0);
+      float glue = (extra_space + i) / (i + 1.0f);
       float gap =
         sizes[(spreading[i])].natural_size - sizes[(spreading[i])].minimum_size;
 
@@ -654,9 +654,9 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
   sizes = g_newa (RequestedSize, nvis_children);
 
   if (priv->orientation == CLUTTER_ORIENTATION_VERTICAL)
-    size = box->y2 - box->y1 - (nvis_children - 1) * priv->spacing;
+    size = (int) (box->y2 - box->y1 - (nvis_children - 1) * priv->spacing);
   else
-    size = box->x2 - box->x1 - (nvis_children - 1) * priv->spacing;
+    size = (int) (box->x2 - box->x1 - (nvis_children - 1) * priv->spacing);
 
   actor = CLUTTER_ACTOR (container);
 
@@ -710,7 +710,7 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
                    ? box->x2 - box->x1
                    : box->y2 - box->y1);
 
-      size -= sizes[i].minimum_size;
+      size -= (int) sizes[i].minimum_size;
 
       sizes[i].actor = child;
 
@@ -723,9 +723,9 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
        * minimum sizes for children that are not going to fill
        */
       if (priv->orientation == CLUTTER_ORIENTATION_VERTICAL)
-        size = box->y2 - box->y1 - (nvis_children - 1) * priv->spacing;
+        size = (int) (box->y2 - box->y1 - (nvis_children - 1) * priv->spacing);
       else
-        size = box->x2 - box->x1 - (nvis_children - 1) * priv->spacing;
+        size = (int) (box->x2 - box->x1 - (nvis_children - 1) * priv->spacing);
 
       extra = size / nvis_children;
       n_extra_widgets = size % nvis_children;
@@ -763,14 +763,14 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
   if (priv->orientation == CLUTTER_ORIENTATION_VERTICAL)
     {
       child_allocation.x1 = box->x1;
-      child_allocation.x2 = MAX (1.0, box->x2);
-      y = box->y1;
+      child_allocation.x2 = MAX (1.0f, box->x2);
+      y = (int) box->y1;
     }
   else
     {
       child_allocation.y1 = box->y1;
-      child_allocation.y2 = MAX (1.0, box->y2);
-      x = box->x1;
+      child_allocation.y2 = MAX (1.0f, box->y2);
+      x = (int) box->x1;
     }
 
   i = 0;
@@ -814,7 +814,7 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
           if (clutter_actor_needs_expand (child, priv->orientation))
             {
               child_allocation.y1 = y;
-              child_allocation.y2 = child_allocation.y1 + MAX (1.0, child_size);
+              child_allocation.y2 = child_allocation.y1 + MAX (1.0f, child_size);
             }
           else
             {
@@ -822,14 +822,14 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
               child_allocation.y2 = child_allocation.y1 + sizes[i].minimum_size;
             }
 
-          y += child_size + priv->spacing;
+          y += (int) (child_size + priv->spacing);
         }
       else /* CLUTTER_ORIENTATION_HORIZONTAL */
         {
           if (clutter_actor_needs_expand (child, priv->orientation))
             {
               child_allocation.x1 = x;
-              child_allocation.x2 = child_allocation.x1 + MAX (1.0, child_size);
+              child_allocation.x2 = child_allocation.x1 + MAX (1.0f, child_size);
             }
           else
             {
@@ -837,7 +837,7 @@ clutter_box_layout_allocate (ClutterLayoutManager   *layout,
               child_allocation.x2 = child_allocation.x1 + sizes[i].minimum_size;
             }
 
-          x += child_size + priv->spacing;
+          x += (int) (child_size + priv->spacing);
 
           if (is_rtl)
             {
