+++ clutter/clutter/clutter-deform-effect.c
@@ -27,7 +27,7 @@
 
 /**
  * ClutterDeformEffect:
- * 
+ *
  * A base class for effects deforming the geometry of an actor
  *
  * #ClutterDeformEffect is an abstract class providing all the plumbing
@@ -237,7 +237,8 @@ clutter_deform_effect_paint_target (ClutterOffscreenEffect *effect,
               vertex.y = height * vertex.ty;
               vertex.z = 0.0f;
 
-              cogl_color_init_from_4f (&vertex.color, 1.0, 1.0, 1.0, opacity / 255.0);
+              cogl_color_init_from_4f (&vertex.color,
+                                       1.0f, 1.0f, 1.0f, opacity / 255.0f);
 
               clutter_deform_effect_deform_vertex (self,
                                                    width, height,
@@ -250,10 +251,10 @@ clutter_deform_effect_paint_target (ClutterOffscreenEffect *effect,
               vertex_out->z = vertex.z;
               vertex_out->s = vertex.tx;
               vertex_out->t = vertex.ty;
-              vertex_out->r = cogl_color_get_red (&vertex.color) * 255.0;
-              vertex_out->g = cogl_color_get_green (&vertex.color) * 255.0;
-              vertex_out->b = cogl_color_get_blue (&vertex.color) * 255.0;
-              vertex_out->a = cogl_color_get_alpha (&vertex.color) * 255.0;
+              vertex_out->r = (uint8_t) (cogl_color_get_red (&vertex.color) * 255.0f);
+              vertex_out->g = (uint8_t) (cogl_color_get_green (&vertex.color) * 255.0f);
+              vertex_out->b = (uint8_t) (cogl_color_get_blue (&vertex.color) * 255.0f);
+              vertex_out->a = (uint8_t) (cogl_color_get_alpha (&vertex.color) * 255.0f);
             }
         }
